+++ Portal/PortalNodeContainerFacade.php
@@ -8,7 +8,6 @@ use Heptacom\HeptaConnect\Core\Portal\Contract\PortalNodeContainerFacadeContract
 use Heptacom\HeptaConnect\Core\Portal\Exception\ServiceNotFoundException;
 use Heptacom\HeptaConnect\Core\Support\HttpMiddlewareCollector;
 use Heptacom\HeptaConnect\Portal\Base\Parallelization\Support\ResourceLockFacade;
-use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalStorageInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Portal\Base\Web\Http\Contract\HttpClientContract;
@@ -40,11 +39,6 @@ final class PortalNodeContainerFacade extends PortalNodeContainerFacadeContract
                 throw new ServiceNotFoundException($returnType->getName(), 1666461305, $callException);
             }
         }
-    }
-
-    public function getPortal(): PortalContract
-    {
-        return $this->get(PortalContract::class);
     }
 
     public function getPortalNodeKey(): PortalNodeKeyInterface